Removes compiled files from project.
These files shouldn't have been included in the project in the first place.
When we deploy or when a different dev pulls down the project these files will be created by running the appropriate gulp command.

I.e. If we're deploying to production, we don't want out dev code (source maps, unminified code) already in the webroot, we want the production version (minified, no maps, etc).