Refactors the ManagesSignups and NotifySlackChannel classes to propertly pass initee instances.
This is correcting some mistakes I made while trying to better learn unit testing. I got the unit
tests to work, but got so caught up in the unit tests that I forgot to make sure it was working for
the system I was creating. I don't know if these are necessarily the _best_ fixes for these classes
but at the very least it gets the api back to functional so I can finish it out and come back to
refactor later.